Add a sort module
It's now possible for a theme using the sort extra to let users
configure the sort order.

Currently supported criteria are group, file, line and access.

The configuration is done with a `sort` context key. It's an array of
strings, representing a sort criteria plus an optional sort order. The
sort order is determined if the last character is either `>` or `<`
(respectively desc and asc).

Example:

    {
      "sort": [
        "access",
        "line>",
        "group",
        "file"
      ]
    }

* See SassDoc/sassdoc#239